Thread state tracking improved
But not perfect.

The listener thread starts suspended.
You can run it, and it works.
Once you suspend it again, we feed it a null byte to push past the accept
method and allow the thread state to suspend.

At this point the listener is broken.
The thread can be resumed, but the listener will not accept any data.

Connecting clients will hang or time out.