fixed persistance of listener
Changed the <EOF> indicator to a newline, so that most every echo sent will
terminate correctly.

When escaping the accept, I send a newline, this keeps the listener working
when the thread id resumed.

The listener wasn't broken during thread suspention, it was because I did not
send a message terminator when attempting to release the accept.

When the receive gets bytes, and they do not contain a terminator, the
inner loop is infinite: the variable, bytes is always that input, no matter
what else is sent into the socket.