fixed chance of frozen listener
The way the example listener was organized, if whatever was set to be the end
of message bytes was not received, the listener would lock up, and not
accept any more lines.

I don't see the advantage of this, except if your client were always
connected and always sending according to protocol without error.

I took out the receive loop. Each termination of the connection sends the
data to be processed.

The thread can be suspended and resumed at will.